requirements.txt
tensorflow
google-cloud-bigquery
google-cloud-pubsub
numpy